Shuffled some more files related to just demo datasets.
