+++ evalml/demos/diabetes.py
@@ -3,7 +3,7 @@
 import woodwork as ww
 
 import evalml
-from evalml.preprocessing import load_data
+from evalml.utils.file_utils import load_data
 
 
 def load_diabetes():
