+++ evalml/demos/weather.py
@@ -2,7 +2,7 @@
 import pandas as pd
 
 import evalml
-from evalml.preprocessing import load_data
+from evalml.utils.file_utils import load_data
 from evalml.utils import infer_feature_types
