+++ evalml/preprocessing/__init__.py
@@ -1,8 +1,6 @@
 """Preprocessing utilities."""
 from .utils import (
-    load_data,
     split_data,
-    number_of_features,
-    target_distribution,
 )
+from ..utils.file_utils import load_data, number_of_features, target_distribution
 from .data_splitters import NoSplit, TrainingValidationSplit, TimeSeriesSplit
