+++ evalml/preprocessing/utils.py
@@ -39,41 +39,6 @@ from evalml.utils import import_or_raise, infer_feature_types
 logger = logging.getLogger(__name__)
 
 
-def load_data(path, index, target, n_rows=None, drop=None, verbose=True, **kwargs):
-    """Load features and target from file.
-
-    Args:
-        path (str): Path to file or a http/ftp/s3 URL.
-        index (str): Column for index.
-        target (str): Column for target.
-        n_rows (int): Number of rows to return. Defaults to None.
-        drop (list): List of columns to drop. Defaults to None.
-        verbose (bool): If True, prints information about features and target. Defaults to True.
-        **kwargs: Other keyword arguments that should be passed to panda's `read_csv` method.
-
-    Returns:
-        pd.DataFrame, pd.Series: Features matrix and target.
-    """
-    feature_matrix = pd.read_csv(path, index_col=index, nrows=n_rows, **kwargs)
-
-    targets = [target] + (drop or [])
-    y = feature_matrix[target]
-    X = feature_matrix.drop(columns=targets)
-
-    if verbose:
-        # number of features
-        print(number_of_features(X.dtypes), end="\n\n")
-
-        # number of total training examples
-        info = "Number of training examples: {}"
-        print(info.format(len(X)), end="\n")
-
-        # target distribution
-        print(target_distribution(y))
-
-    return infer_feature_types(X), infer_feature_types(y)
-
-
 def split_data(
     X, y, problem_type, problem_configuration=None, test_size=0.2, random_seed=0
 ):
@@ -142,72 +107,6 @@ def split_data(
     y_test = y.ww.iloc[test]
 
     return X_train, X_test, y_train, y_test
-
-
-def number_of_features(dtypes):
-    """Get the number of features of each specific dtype in a DataFrame.
-
-    Args:
-        dtypes (pd.Series): DataFrame.dtypes to get the number of features for.
-
-    Returns:
-        pd.Series: dtypes and the number of features for each input type.
-
-    Example:
-        >>> X = pd.DataFrame()
-        >>> X["integers"] = [i for i in range(10)]
-        >>> X["floats"] = [float(i) for i in range(10)]
-        >>> X["strings"] = [str(i) for i in range(10)]
-        >>> X["booleans"] = [bool(i%2) for i in range(10)]
-
-        Lists the number of columns corresponding to each dtype.
-
-        >>> number_of_features(X.dtypes)
-                     Number of Features
-        Boolean                       1
-        Categorical                   1
-        Numeric                       2
-    """
-    dtype_to_vtype = {
-        "bool": "Boolean",
-        "int32": "Numeric",
-        "int64": "Numeric",
-        "float64": "Numeric",
-        "object": "Categorical",
-        "datetime64[ns]": "Datetime",
-    }
-
-    vtypes = dtypes.astype(str).map(dtype_to_vtype).value_counts()
-    return vtypes.sort_index().to_frame("Number of Features")
-
-
-def target_distribution(targets):
-    """Get the target distributions.
-
-    Args:
-        targets (pd.Series): Target data.
-
-    Returns:
-        pd.Series: Target data and their frequency distribution as percentages.
-
-    Examples:
-        >>> y = pd.Series([1, 2, 4, 1, 3, 3, 1, 2])
-        >>> target_distribution(y)
-        Targets
-        1    37.50%
-        2    25.00%
-        3    25.00%
-        4    12.50%
-        dtype: object
-        >>> y = pd.Series([True, False, False, False, True])
-        >>> target_distribution(y)
-        Targets
-        False    60.00%
-        True     40.00%
-        dtype: object
-    """
-    distribution = targets.value_counts() / len(targets)
-    return distribution.mul(100).apply("{:.2f}%".format).rename_axis("Targets")
 
 
 def _get_preprocessing_components(
